requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1
python-dotenv==0.18.0
nltk==3.6.2
